패키지 분리 & 경보, PCR정보 API 호출
