+++ src/main/java/com/travelplan/domain/log/domain/Log.java
@@ -1,0 +1,4 @@
+package com.travelplan.domain.log.domain;
+
+public class Log {
+}
