+++ src/main/java/com/travelplan/domain/travel/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package com.travelplan.domain.travel.repository;
+
+import com.travelplan.domain.user.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User,Integer> {
+    Optional<User> findByEmail(String email);
+}
