--- src/main/java/com/travelplan/dto/WarningDto.java
@@ -1,22 +1,0 @@
-package com.travelplan.dto;
-
-import lombok.*;
-
-import java.util.List;
-
-@Getter
-@Setter
-public class WarningDto {
-
-    List<WarningContentData> data;
-
-    @Data
-    static class WarningContentData {
-        private String alarm_lvl;
-        private String country_eng_nm;
-        private String country_iso_alp2;
-        private String country_nm;
-        private String written_dt;
-    }
-
-}
