+++ src/main/java/com/travelplan/global/config/auth/CustomOAuth2UserService.java
@@ -2,8 +2,8 @@ package com.travelplan.global.config.auth;
 
 import com.travelplan.global.config.auth.dto.OAuthAttributes;
 import com.travelplan.global.config.auth.dto.SessionUser;
-import com.travelplan.travel.domain.User;
-import com.travelplan.travel.repository.UserRepository;
+import com.travelplan.domain.user.domain.User;
+import com.travelplan.domain.travel.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
