+++ src/main/java/com/travelplan/global/entity/code/JoinStatus.java
@@ -1,0 +1,5 @@
+package com.travelplan.global.entity.code;
+
+public enum JoinStatus {
+
+}
