+++ src/main/java/com/travelplan/global/entity/code/UserRole.java
@@ -1,0 +1,12 @@
+package com.travelplan.global.entity.code;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Getter
+public enum UserRole {
+    USER("ROLE_USER"), ADMIN("ROLE_ADMIN");
+    private final String key;
+}
+
