+++ src/main/java/com/travelplan/global/entity/code/Vehicle.java
@@ -1,0 +1,5 @@
+package com.travelplan.global.entity.code;
+
+public enum Vehicle {
+    SUBWAY,CAR
+}
