--- src/main/java/com/travelplan/travel/domain/Log.java
@@ -1,4 +1,0 @@
-package com.travelplan.travel.domain;
-
-public class Log {
-}
