--- src/main/java/com/travelplan/travel/domain/code/JoinStatus.java
@@ -1,5 +1,0 @@
-package com.travelplan.travel.domain.code;
-
-public enum JoinStatus {
-
-}
