--- src/main/java/com/travelplan/travel/domain/code/MemberRole.java
@@ -1,5 +1,0 @@
-package com.travelplan.travel.domain.code;
-
-public enum MemberRole {
-
-}
