--- src/main/java/com/travelplan/travel/domain/code/UserRole.java
@@ -1,12 +1,0 @@
-package com.travelplan.travel.domain.code;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Getter
-public enum UserRole {
-    USER("ROLE_USER"), ADMIN("ROLE_ADMIN");
-    private final String key;
-}
-
