--- src/main/java/com/travelplan/travel/domain/code/Vehicle.java
@@ -1,5 +1,0 @@
-package com.travelplan.travel.domain.code;
-
-public enum Vehicle {
-    SUBWAY,CAR
-}
