--- src/main/java/com/travelplan/travel/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package com.travelplan.travel.repository;
-
-import com.travelplan.travel.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User,Integer> {
-    Optional<User> findByEmail(String email);
-}
